dh-ag95-gripper/dh-ag95-gripper-backend/requirements.txt
pyserial == 3.5
minimalmodbus == 2.1.1
